# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectorKind": {
        "enum": [
          "DRIVE",
          "GMAIL"
        ],
        "title": "ConnectorKind",
        "type": "string"
      },
      "ConnectorRead": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ConnectorKind"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectorStatus"
          }
        },
        "required": [
          "id",
          "kind",
          "status"
        ],
        "title": "ConnectorRead",
        "type": "object"
      },
...
      "ConnectorStatus": {
        "enum": [
          "ACTIVE",
          "REAUTH_REQUIRED",
          "ERROR"
        ],
        "title": "ConnectorStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "PublicChatCredential": {
        "bearerFormat": "session credential",
        "description": "Opaque credential issued for one public chat session.",
        "scheme": "bearer",
        "type": "http"
      },
      "StaffSession": {
        "description": "Issued after approved staff OIDC admission; state-changing routes also require CSRF.",
        "in": "cookie",
        "name": "staff_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Enterprise AI Agent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/connectors/{connector_id}/revoke": {
      "post": {
        "operationId": "revoke_api_v1_admin_connectors__connector_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connector Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Revoke",
        "tags": [
          "connectors"
        ],
        "x-security-classification": "administrator"
      }
    },
    "/api/v1/admin/connectors/{kind}/authorize": {
      "get": {
        "operationId": "authorize_api_v1_admin_connectors__kind__authorize_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConnectorKind"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Authorize",
        "tags": [
          "connectors"
        ],
        "x-security-classification": "administrator"
      }
    },
    "/api/v1/admin/connectors/{kind}/callback": {
      "get": {
        "operationId": "connector_callback_api_v1_admin_connectors__kind__callback_get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConnectorKind"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Connector Callback",
        "tags": [
          "connectors"
        ],
        "x-security-classification": "administrator"
      }
    },
...
  }
}